Feat: implemented and computed a lookup table for possible_keys() for the reduced problem (num_digits=4, num_colors=6).
The original problem (num_digits=5, num_colors=8), i.e. the Mastermind I
played as a child, turned out to be so expensive that precomputing
possible_keys() would require a > 100 GB lookup table. Without a lookup
table the program, while correct, is unusably slow.

It turns out Donald E. Knuth in the first paper on the topic ("The
Computer as Master Mind", 1976) considered the reduced version
(num_digits=5, num_colors=8). This is much more tractable; I switch to
it.